+++ app/src/main/java/it/dhd/bcrmanager/observer/ContactObserver.java
@@ -14,7 +14,6 @@ import android.widget.Toast;
 import java.io.File;
 import java.util.List;
 
-import it.dhd.bcrmanager.R;
 import it.dhd.bcrmanager.drawable.LetterTileDrawable;
 import it.dhd.bcrmanager.objects.ContactItem;
 import it.dhd.bcrmanager.utils.FileUtils;
@@ -44,11 +43,12 @@ public class ContactObserver extends ContentObserver {
         this.context = context;
     }
 
+
+
     @Override
     public void onChange(boolean selfChange, Uri uri) {
 
-
-        // Check
+        super.onChange(selfChange, uri);
         if (!new File(context.getFilesDir(), FileUtils.STORED_CONTACTS).exists()) return;
         List<ContactItem> contacts = FileUtils.loadObjectList(context, FileUtils.STORED_CONTACTS, ContactItem.class);
         StringBuilder changedContacts = new StringBuilder();
@@ -56,7 +56,7 @@ public class ContactObserver extends ContentObserver {
         if (contacts.size() == 0) return;
         boolean hasChanged = false;
         if (!TextUtils.isEmpty(contactNumber)) {
-            Log.d("ContactObserver", "QUERY SINGLE NUMBER: " + contactNumber);
+            Log.d("ContactObserver", "QUERY SINGLE NUMBER");
             // Query for a single number
             ContactItem contactFound = null;
             for (ContactItem contact : contacts) {
@@ -69,7 +69,7 @@ public class ContactObserver extends ContentObserver {
             if (contactFound == null) return;
             Cursor cursor = context.getContentResolver().query(
                     Uri.withAppendedPath(ContactsContract.PhoneLookup.CONTENT_FILTER_URI, Uri.encode(contactNumber)),
-                    new String[] {
+                    new String[]{
                             ContactsContract.Contacts._ID,
                             ContactsContract.Contacts.DISPLAY_NAME,
                             ContactsContract.Contacts.PHOTO_URI,
@@ -109,10 +109,10 @@ public class ContactObserver extends ContentObserver {
                 if (!TextUtils.isEmpty(cursor.getString(3))) {
                     Cursor phones = context.getContentResolver().query(
                             ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
-                            new String[] {
+                            new String[]{
                                     ContactsContract.CommonDataKinds.Phone.NUMBER,
                                     ContactsContract.CommonDataKinds.Phone.TYPE,
-                                    ContactsContract.CommonDataKinds.Phone.LABEL } ,
+                                    ContactsContract.CommonDataKinds.Phone.LABEL},
                             ContactsContract.CommonDataKinds.Phone.CONTACT_ID + " = " + cursor.getLong(0),
                             null,
                             null);
@@ -126,7 +126,8 @@ public class ContactObserver extends ContentObserver {
                             String numberLabel;
                             if (phones.getInt(1) == ContactsContract.CommonDataKinds.Phone.TYPE_CUSTOM)
                                 numberLabel = phones.getString(2);
-                            else numberLabel = (String) ContactsContract.CommonDataKinds.Phone.getTypeLabel(context.getResources(), phones.getInt(1), "");
+                            else
+                                numberLabel = (String) ContactsContract.CommonDataKinds.Phone.getTypeLabel(context.getResources(), phones.getInt(1), "");
                             if (!TextUtils.equals(contactFound.getNumberLabel(), numberLabel)) {
                                 hasChanged = true;
                                 contactFound.setNumberLabel(numberLabel);
@@ -134,7 +135,7 @@ public class ContactObserver extends ContentObserver {
                             break;
                         }
                     }
-                    if(phones != null && !phones.isClosed()) {
+                    if (phones != null && !phones.isClosed()) {
                         phones.close();
                     }
                 }
@@ -144,121 +145,32 @@ public class ContactObserver extends ContentObserver {
                 contactFound.resetContact();
             }
             if (hasChanged) {
-                count ++;
+                count++;
                 changedContacts.append("N:").append(contactFound.getPhoneNumber()).append(";");
             }
-            if(cursor != null && !cursor.isClosed()) {
+            if (cursor != null && !cursor.isClosed()) {
                 cursor.close();
-            }
-        } else {
-            // Query for all contacts
-            for (ContactItem contact : contacts) {
-                Cursor cursor = context.getContentResolver().query(
-                        Uri.withAppendedPath(ContactsContract.PhoneLookup.CONTENT_FILTER_URI, Uri.encode(contact.getPhoneNumber())),
-                        new String[]{
-                                ContactsContract.PhoneLookup._ID,
-                                ContactsContract.PhoneLookup.DISPLAY_NAME,
-                                ContactsContract.PhoneLookup.PHOTO_URI,
-                                ContactsContract.PhoneLookup.LOOKUP_KEY
-                        },
-                        null,
-                        null,
-                        null
-                );
-                if (cursor != null && cursor.moveToFirst()) {
-                    do {
-                        if (cursor.getString(1).isEmpty()) {
-                            // Probably contact was deleted
-                            hasChanged = true;
-                            contact.resetContact();
-                        } else {
-                            // Name changed
-                            if (!TextUtils.equals(contact.getContactName(), cursor.getString(1))) {
-                                hasChanged = true;
-                                contact.setContactName(cursor.getString(1));
-                                if (!contact.isContactSaved()) {
-                                    contact.setContactSaved(true);
-                                    contact.setContactType(LetterTileDrawable.TYPE_PERSON);
-                                }
-                            }
-                        }
-                        // Check contact Image
-                        if (!UriUtils.areEqual(contact.getContactImage(), UriUtils.parseUriOrNull(cursor.getString(2)))) {
-                            hasChanged = true;
-                            contact.setContactImage(UriUtils.parseUriOrNull(cursor.getString(2)));
-                        }
-                        // Check contact LookupKey
-                        if (!TextUtils.equals(contact.getLookupKey(), cursor.getString(3))) {
-                            hasChanged = true;
-                            contact.setLookupKey(cursor.getString(3));
-                        }
-                        // Check contact ID
-                        if (contact.getContactId() != cursor.getLong(0)) {
-                            hasChanged = true;
-                            contact.setContactId(cursor.getLong(0));
-                        }
-                        // Check Number Type and Label
-                        if (!TextUtils.isEmpty(cursor.getString(3))) {
-                            Cursor phones = context.getContentResolver().query(
-                                    ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
-                                    new String[] {
-                                            ContactsContract.CommonDataKinds.Phone.NUMBER,
-                                            ContactsContract.CommonDataKinds.Phone.TYPE,
-                                            ContactsContract.CommonDataKinds.Phone.LABEL } ,
-                                    ContactsContract.CommonDataKinds.Phone.CONTACT_ID + " = " + cursor.getLong(0),
-                                    null,
-                                    null);
-                            while (phones != null && phones.moveToNext()) {
-                                if (PhoneNumberUtils.compare(contact.getPhoneNumber(), phones.getString(0))) {
-                                    // Number Type changed
-                                    if (contact.getNumberType() != phones.getInt(1)) {
-                                        hasChanged = true;
-                                        contact.setContactType(phones.getInt(1));
-                                    }
-                                    String numberLabel;
-                                    if (phones.getInt(1) == ContactsContract.CommonDataKinds.Phone.TYPE_CUSTOM)
-                                        numberLabel = phones.getString(2);
-                                    else numberLabel = (String) ContactsContract.CommonDataKinds.Phone.getTypeLabel(context.getResources(), phones.getInt(1), "");
-                                    if (TextUtils.equals(contact.getNumberLabel(), numberLabel)) {
-                                        hasChanged = true;
-                                        contact.setNumberLabel(numberLabel);
-                                    }
-                                    break;
-                                }
-                            }
-                            if(phones != null && !phones.isClosed()) {
-                                phones.close();
-                            }
-                        }
-                        if (hasChanged) {
-                            count ++;
-                            changedContacts.append("N:").append(contact.getPhoneNumber()).append(";");
-                        }
-                    } while (cursor.moveToNext());
-                }
-                if(cursor != null && !cursor.isClosed()) {
-                    cursor.close();
-                }
             }
         }
         if (count == 0) {
             if (dataUpdateListener != null) {
-                dataUpdateListener.onDataUpdate("", null);
+                dataUpdateListener.onDataUpdate(false);
             }
             return;
         }
         FileUtils.saveObjectList(context, contacts, FileUtils.STORED_CONTACTS, ContactItem.class);
-        String formattedString = context.getResources().getQuantityString(R.plurals.contact_changed, count, count);
-        showToast(formattedString);
+        if (!TextUtils.isEmpty(contactNumber)) {
+            showToast("MAMMT");
+        }
 
         if (dataUpdateListener != null) {
-            dataUpdateListener.onDataUpdate(changedContacts.toString(), contacts);
+            dataUpdateListener.onDataUpdate(true);
         }
 
     }
 
     public interface DataUpdateListener {
-        void onDataUpdate(String newData, List<ContactItem> newContacts);
+        void onDataUpdate(boolean requireUpdate);
     }
 
     private void showToast(String message) {