+++ app/src/main/java/it/dhd/bcrmanager/ui/dialogs/DeleteDialog.java
@@ -2,7 +2,6 @@ package it.dhd.bcrmanager.ui.dialogs;
 
 import android.annotation.SuppressLint;
 import android.app.Dialog;
-import android.net.Uri;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -20,14 +19,11 @@ import it.dhd.bcrmanager.R;
 import it.dhd.bcrmanager.databinding.DeleteDialogBinding;
 import it.dhd.bcrmanager.objects.CallLogItem;
 import it.dhd.bcrmanager.utils.CircleTransform;
-import it.dhd.bcrmanager.utils.FileUtils;
 import it.dhd.bcrmanager.utils.PreferenceUtils;
-import it.dhd.bcrmanager.utils.UriUtils;
 
 public class DeleteDialog extends AppCompatDialogFragment {
 
     private final CallLogItem itemToDelete;
-    private DeleteDialogBinding binding;
 
     private final onItemDeletedListener onItemDeleted;
 
@@ -46,68 +42,48 @@ public class DeleteDialog extends AppCompatDialogFragment {
     @Override
     public Dialog onCreateDialog(Bundle savedInstanceState) {
         MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(requireActivity());
-        binding = DeleteDialogBinding.inflate(requireActivity().getLayoutInflater());
+        it.dhd.bcrmanager.databinding.DeleteDialogBinding binding = DeleteDialogBinding.inflate(requireActivity().getLayoutInflater());
 
         builder.setView(binding.getRoot());
 
-        assert getArguments() != null;
-        int nSim = getArguments().getInt("nSim");
+        binding.itemEntry.setShowIcon(PreferenceUtils.showIcon());
+        binding.itemEntry.setShowLabel(PreferenceUtils.showLabel());
+        binding.itemEntry.setShowLabel(PreferenceUtils.showLabel());
+        binding.itemEntry.setCallLogItem(itemToDelete);
+        binding.itemEntry.actionPlay.setVisibility(View.GONE);
+        binding.itemEntry.expandingLayout.setVisibility(View.GONE);
 
-        String phoneNumber = itemToDelete.getNumber();
-        if (phoneNumber != null && !phoneNumber.isEmpty()) {
-            if (itemToDelete.getContactIcon() != null) {
-                Picasso.get().load(itemToDelete.getContactIcon()).error(R.drawable.ic_default_contact).placeholder(R.drawable.ic_default_contact).transform(new CircleTransform()).into(binding.contactIconDelete);
-            } else if (PreferenceUtils.showTiles()) {
-                binding.contactIconDelete.setImageDrawable(itemToDelete.getContactDrawable(requireContext()));
-            } else {
-                binding.contactIconDelete.setImageResource(R.drawable.ic_default_contact);
-            }
-            binding.contactNameDelete.setText(itemToDelete.getContactName());
-        }
-
-        if (nSim<=1) {
-            binding.dividerSimDelete.setVisibility(View.GONE);
-            binding.simSlotDelete.setVisibility(View.GONE);
-            binding.dividerDateDelete.setVisibility(View.GONE);
+        if (itemToDelete.getContactIcon() != null) {
+            Picasso.get().load(itemToDelete.getContactIcon()).error(R.drawable.ic_default_contact).placeholder(R.drawable.ic_default_contact).transform(new CircleTransform()).into(binding.itemEntry.contactIcon);
+        } else if (PreferenceUtils.showTiles()) {
+            binding.itemEntry.contactIcon.setImageDrawable(itemToDelete.getContactDrawable(requireContext()));
         } else {
-            binding.simSlotDelete.setText(itemToDelete.getSimSlot().toString());
+            binding.itemEntry.contactIcon.setImageResource(R.drawable.ic_default_contact);
         }
 
-        if(itemToDelete.getDirection().contains("in")) binding.callIcon.setImageResource(R.drawable.ic_in);
-        else binding.callIcon.setImageResource(R.drawable.ic_out);
+        switch (itemToDelete.getDirection()) {
+            case "out" -> binding.itemEntry.callIcon.setImageResource(R.drawable.ic_out);
+            case "conference" -> binding.itemEntry.callIcon.setImageResource(R.drawable.ic_conference);
+            default -> binding.itemEntry.callIcon.setImageResource(R.drawable.ic_in);
+        }
 
-        binding.dateDelete.setText(itemToDelete.getFormattedTimestamp(getString(R.string.today), getString(R.string.yesterday)));
+        binding.itemEntry.date.setText(itemToDelete.getFormattedTimestamp(getString(R.string.today), getString(R.string.yesterday)));
 
-        binding.durationDelete.setText(itemToDelete.getFormattedDuration(getString(R.string.format_sec), getString(R.string.format_min)));
-
-
-        // Set click listener for the expand button
-        //holder.expandButton.setOnClickListener(v -> toggleExpansion(holder.expandingLayout));
         binding.btnNegative.setOnClickListener(v -> {
             // Dismiss the alert dialog
             Objects.requireNonNull(getDialog()).cancel();
         });
+
         binding.btnPositive.setOnClickListener(v -> {
             // Dismiss the alert dialog
             MaterialAlertDialogBuilder mBuilder = new MaterialAlertDialogBuilder(requireActivity());
             mBuilder.setTitle("Are you sure?")
-                    .setPositiveButton("Yes", (dialog, which) -> {
+                    .setPositiveButton(android.R.string.yes, (dialog, which) -> {
                         // Dismiss the alert dialog
+                        dialog.dismiss();
                         Objects.requireNonNull(getDialog()).dismiss();
                         // Call the method in MainActivity to delete this item
-
                         onItemDeleted.onItemDeleted();
-
-                        Uri fileUri, audioUri;
-                        fileUri = Uri.parse(itemToDelete.getFilePath());
-                        audioUri = Uri.parse(itemToDelete.getAudioFilePath());
-
-                        if (UriUtils.areEqual(fileUri, audioUri)) {
-                            FileUtils.deleteFileUri(requireContext(), audioUri);
-                        } else {
-                            FileUtils.deleteFileUri(requireContext(), fileUri);
-                            FileUtils.deleteFileUri(requireContext(), audioUri);
-                        }
 
                     });
             mBuilder.show();
@@ -120,10 +96,8 @@ public class DeleteDialog extends AppCompatDialogFragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         DeleteDialogBinding binding = DeleteDialogBinding.inflate(inflater, container, false);
-        assert getArguments() != null;
-        String message = getArguments().getString("message");
 
-        Objects.requireNonNull(getDialog()).setTitle(message);
+        Objects.requireNonNull(getDialog()).setTitle(requireContext().getString(R.string.delete_dialog_message));
 
         return binding.getRoot();
     }