+++ app/src/main/java/it/dhd/bcrmanager/ui/fragments/settings/Settings.java
@@ -3,9 +3,11 @@ package it.dhd.bcrmanager.ui.fragments.settings;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+import android.text.TextUtils;
 
 import androidx.documentfile.provider.DocumentFile;
 import androidx.fragment.app.FragmentTransaction;
+import androidx.lifecycle.ViewModelProvider;
 import androidx.preference.DropDownPreference;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
@@ -16,17 +18,22 @@ import java.util.Objects;
 import it.dhd.bcrmanager.BuildConfig;
 import it.dhd.bcrmanager.MainActivity;
 import it.dhd.bcrmanager.R;
-import it.dhd.bcrmanager.ui.fragments.ItemSettings;
+import it.dhd.bcrmanager.ui.fragments.settingscontainers.BottomPlayerSettings;
+import it.dhd.bcrmanager.ui.fragments.settingscontainers.ItemSettings;
 import it.dhd.bcrmanager.utils.FileUtils;
 import it.dhd.bcrmanager.utils.PreferenceUtils;
+import it.dhd.bcrmanager.viewmodel.FileViewModel;
 
 public class Settings extends  PreferenceFragmentCompat {
 
     private Preference mDirPref;
+    private FileViewModel fileModel;
 
     @Override
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
         setPreferencesFromResource(R.xml.settings, rootKey);
+        requireActivity().setTitle(R.string.settings);
+        fileModel = new ViewModelProvider(requireActivity()).get(FileViewModel.class);
         mDirPref = findPreference("bcr_directory");
         Uri treeUri = Uri.parse(PreferenceUtils.getStoredFolderFromPreference());
         DocumentFile pickedDir = DocumentFile.fromTreeUri(requireContext(), treeUri);
@@ -44,6 +51,19 @@ public class Settings extends  PreferenceFragmentCompat {
                         .beginTransaction()
                         .replace(R.id.frame_layout, new ItemSettings(), ItemSettings.class.getSimpleName())
                         .addToBackStack(ItemSettings.class.getSimpleName())
+                        .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_OPEN)
+                        .commit();
+                return true;
+            });
+        }
+
+        Preference playerPref = findPreference("bottom_player");
+        if (playerPref != null) {
+            playerPref.setOnPreferenceClickListener(pref -> {
+                requireActivity().getSupportFragmentManager()
+                        .beginTransaction()
+                        .replace(R.id.frame_layout, new BottomPlayerSettings(), BottomPlayerSettings.class.getSimpleName())
+                        .addToBackStack(BottomPlayerSettings.class.getSimpleName())
                         .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_OPEN)
                         .commit();
                 return true;
@@ -102,5 +122,8 @@ public class Settings extends  PreferenceFragmentCompat {
         Uri treeUri = Uri.parse(PreferenceUtils.getStoredFolderFromPreference());
         DocumentFile pickedDir = DocumentFile.fromTreeUri(requireContext(), treeUri);
         mDirPref.setSummary(FileUtils.getPathFromUri(requireContext(), Objects.requireNonNull(pickedDir).getUri()));
+        if (!TextUtils.equals(PreferenceUtils.getLatestFolder(), PreferenceUtils.getStoredFolderFromPreference())) {
+            fileModel.fetchData(requireContext(), null);
+        }
     }
 }